Version June 18, 2021

+++ data_engineering/transform/database_normalization.py
@@ -21,7 +21,8 @@ def calling_transformed_files(path, csv_from_path=['npi', 'npri', 'tri']):
         df = pd.concat([df, df_u],
                         axis=0, ignore_index=True)
         del df_u
-        os.remove(f'{path}/{file}.csv')
+        if not file == 'National_to_generic_transfer':
+            os.remove(f'{path}/{file}.csv')
 
     return df
 
@@ -42,10 +43,10 @@ def database_normalization():
     country_to_ics = {'USA': 'USA_NAICS', 'AUS': 'ANZSIC', 'CAN': 'CAN_NAICS'}
     prtr['industry_classification_system'] = prtr.country.apply(lambda x: country_to_ics[x])
     prtr.drop(columns=['cas_number', 'national_substance_name'], inplace=True)
-
-    # Creating 'national_facility_and_sector_id'
-    grouping = ['national_facility_id', 'national_sector_code', 'prtr_system']
-    prtr['national_facility_and_sector_id'] = pd.Series(prtr.groupby(grouping).ngroup())
+    prtr.reset_index(inplace=True, drop=True)
+    prtr['transfer_record_id'] = pd.Series(prtr.index.tolist()) + 1
+    prtr.reliability_score = prtr.reliability_score.astype(int)
+    prtr_columns = list(prtr.columns)
 
     # Calling sectors
     sector = calling_transformed_files(f'{dir_path}/output',
@@ -53,7 +54,7 @@ def database_normalization():
 
     # Calling substances
     substance = calling_transformed_files(f'{dir_path}/output',
-                                    csv_from_path=['national_to_generic_substance'])
+                                    csv_from_path=['national_to_generic_substance'])     
 
     # Calling transfer classes
     t_class = calling_transformed_files(f'{dir_path}/../../ancillary',
@@ -63,6 +64,7 @@ def database_normalization():
         t_class.groupby('generic_transfer_class_id')\
             .generic_system_comment\
                 .transform(lambda g: g.fillna(method='bfill'))
+    t_class.national_generic_transfer_class_id = t_class.national_generic_transfer_class_id.astype(int)
 
     # Merging dataframes
     mergings = [[sector, ['national_sector_code', 'industry_classification_system']],
@@ -72,19 +74,38 @@ def database_normalization():
     for merging in mergings:
         prtr = pd.merge(prtr, merging[0], how='left', on=merging[1])
     del mergings
+    prtr.drop_duplicates(keep='first', subset=prtr_columns, inplace=True)
 
+    # Creating 'national_transfer_class_prtr_system_id'
+    grouping = ['national_transfer_class_id', 'prtr_system']
+    prtr['national_transfer_class_prtr_system_id'] = pd.Series(prtr.groupby(grouping).ngroup() + 1)
 
-    # Creating 'national_sector_and_classification_system_id'
-    grouping = ['national_sector_code', 'national_sector_name', 'industry_classification_system']
-    prtr['national_sector_and_classification_system_id'] = pd.Series(prtr.groupby(grouping).ngroup())
+    # Creating 'national_sector_id'
+    grouping = ['national_sector_code', 'industry_classification_system']
+    prtr['national_sector_id'] = pd.Series(prtr.groupby(grouping).ngroup() + 1)
+
+    # Creating 'national_generic_sector_id'
+    grouping = ['national_sector_id', 'generic_sector_code']
+    prtr['national_generic_sector_id'] = pd.Series(prtr.groupby(grouping).ngroup() + 1)
+
+    # Creating 'national_substance_prtr_system_id'
+    grouping = ['national_substance_id', 'prtr_system']
+    prtr['national_substance_prtr_system_id'] = pd.Series(prtr.groupby(grouping).ngroup() + 1)
+
+    # Creating 'national_generic_substance_id'
+    grouping = ['national_substance_prtr_system_id', 'generic_substance_id']
+    prtr['national_generic_substance_id'] = pd.Series(prtr.groupby(grouping).ngroup() + 1)
+
+    # Creating 'national_facility_and_generic_sector_id'
+    grouping = ['national_facility_id', 'national_generic_sector_id']
+    prtr['national_facility_and_generic_sector_id'] = pd.Series(prtr.groupby(grouping).ngroup() + 1)
 
     # Creating individual tables
-    for table, cols in db_normalization.items():
-        df_table = prtr[cols['cols']]
-        df_table = df_table.drop_duplicates(keep='first').reset_index(drop=True)
+    for table, params in db_normalization.items():
+        df_table = prtr[params['cols']]
+        df_table = df_table.drop_duplicates(keep='first', subset=params['key']).reset_index(drop=True)
         df_table.to_csv(f'{dir_path}/output/{table}.csv',
                         index=False, sep=',')
-    
     
 
 if __name__ == '__main__':
